CookieTest: split `parse[_from_headers]()` specific tests to dedicated test class
Note: while I have added a `@covers` tag, I've not made it method specific yet as these tests still need a detailed review for stability, usability and code coverage.
